# docs/swagger.yaml
openapi: 3.0.3
info:
  title: SaaS-on-SaaS Backend API
  description: Multi-tenant SaaS platform with SQL editor capabilities
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /auth/signup/request-otp:
    post:
      tags:
        - Authentication
      summary: Request OTP for signup
      description: Generate and send OTPs to company and admin email addresses
      operationId: requestOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - company_email
                - admin_email
              properties:
                company_email:
                  type: string
                  format: email
                  description: Company email for OTP verification
                  example: contact@acme.com
                admin_email:
                  type: string
                  format: email
                  description: Admin email for OTP verification
                  example: admin@acme.com
      responses:
        '200':
          description: OTP generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change user password
      description: Update password for authenticated user
      operationId: changePassword
      parameters:
        - name: x-session-id
          in: header
          required: true
          description: Session ID from login
          schema:
            type: integer
            example: 12345
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
                - confirm_password
              properties:
                old_password:
                  type: string
                  format: password
                  description: Current password
                  example: oldPassword123
                new_password:
                  type: string
                  format: password
                  description: New password
                  example: newPassword456
                confirm_password:
                  type: string
                  format: password
                  description: Confirm new password
                  example: newPassword456
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Missing fields or passwords don't match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid session or incorrect current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: header
      name: x-session-id
      description: Session ID obtained from login endpoint
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
              example: 400
            message:
              type: string
              example: Username and password are required
            details:
              type: string
              description: Optional error details
...